refactor part 3
